updates (#19)
This is mostly a documentation update to restructure the BLENDER docs. I'm doing this because git commits are very hard to keep track of especially if one needs to update, then one has to update through all the commits and restructure them all.

So this just inlines all the diffs and resulting files so folks can copy/paste them if they need.